fixed: checkActive and register error checking
In checkActive, I totally forgot to check whether a socket to the target
peer can be created in the first place. If it cannot, we can say for
certain that they are inactive.

In register, I added the ability for a peer to be able to register
multiple times without logging in. If they are logged in though, we
require them to logout first.